requirements.txt
flask
flask-mail
python-dotenv
fpdf
